requirements.txt
gitpython==3.1.27
bottle==0.12.23
requests == 2.31.0
faker == 19.6.2
icecream == 2.1.3
bcrypt == 4.1.2
paste == 3.7.1
python-dotenv == 1.0.1